Push footer to bottom of the page
in the previous version, if the screen height is large, the footer section doesn't stick to the bottom of the screen. in
this version, it will do. body min-height is set to 100vh. footer section can grow vertically if container has free space (in other word, if the screen height is large!).